+++ src/dictionary.js
