+++ src/entryProcessor.js
