+++ src/entryProcessor.test.js
