+++ src/literalFinder.js
