+++ src/literalFinder.test.js
