+++ src/templateStripper.js
