+++ src/templateStripper.test.js
