+++ src/token.js
